# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sistema Jurídico API",
    "description": "Documentação oficial da API do LegalTech",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/auth/redefinir-senha": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Redefinir senha com token",
        "description": "Conclui a redefinição de senha utilizando o token recebido por e-mail",
        "operationId": "redefinirSenha",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedefinirSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Senha redefinida com sucesso",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "422": {
            "description": "Política de senha não atendida ou token inválido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "400": {
            "description": "Dados da requisição inválidos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "404": {
            "description": "Token não encontrado ou expirado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/recuperar-senha": {
      "post": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Solicitar recuperação de senha",
        "description": "Inicia o fluxo de recuperação de senha gerando token temporário enviado por e-mail",
        "operationId": "recuperarSenha",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecuperarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Solicitação aceita; se a conta existir, as instruções foram enviadas",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "E-mail inválido ou malformatado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroValidacaoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Regra de negócio violada",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/auth/me/senha": {
      "patch": {
        "tags": [
          "Autenticação"
        ],
        "summary": "Alterar senha autenticada",
        "description": "Altera a senha do usuário logado mediante confirmação da senha atual",
        "operationId": "alterarSenha",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlterarSenhaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Senha atual inválida ou dados incorretos",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "403": {
            "description": "Acesso proibido",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "204": {
            "description": "Senha alterada com sucesso"
          },
          "404": {
            "description": "Usuário não encontrado",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          },
          "422": {
            "description": "Política de senha não atendida",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ErroPadraoDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErroPadraoDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Data e hora em que o erro ocorreu",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Identificador ou classificação do erro",
            "example": "Bad Request"
          },
          "message": {
            "type": "string",
            "description": "Descrição detalhada da mensagem de erro",
            "example": "Dados inválidos fornecidos"
          }
        },
        "description": "Estrutura padrão de resposta para erros da API"
      },
      "CampoErroDTO": {
        "type": "object",
        "properties": {
          "campo": {
            "type": "string",
            "description": "Nome do campo que falhou na validação",
            "example": "email"
          },
          "mensagem": {
            "type": "string",
            "description": "Mensagem descrevendo o motivo da falha",
            "example": "E-mail inválido"
          }
        },
        "description": "Detalhe do erro de validação em um campo específico"
      },
      "ErroValidacaoDTO": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Data e hora em que o erro ocorreu",
            "format": "date-time"
          },
          "status": {
            "type": "integer",
            "description": "Código HTTP do erro",
            "format": "int32",
            "example": 400
          },
          "error": {
            "type": "string",
            "description": "Identificador ou classificação do erro",
            "example": "Validation Error"
          },
          "message": {
            "type": "string",
            "description": "Mensagem geral do erro de validação",
            "example": "Erro de validação nos campos informados."
          },
          "fieldErrors": {
            "type": "array",
            "description": "Lista detalhada dos erros encontrados por campo",
            "items": {
              "$ref": "#/components/schemas/CampoErroDTO"
            }
          }
        },
        "description": "Estrutura de resposta para erros de validação de campos"
      },
...
      "RedefinirSenhaRequest": {
        "required": [
          "novaSenha",
          "token"
        ],
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "novaSenha": {
            "type": "string"
          }
        }
      },
      "RecuperarSenhaRequest": {
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        }
      },
...
      "AlterarSenhaRequest": {
        "required": [
          "novaSenha",
          "senhaAtual"
        ],
        "type": "object",
        "properties": {
          "senhaAtual": {
            "type": "string"
          },
          "novaSenha": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "name": "bearerAuth",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}